requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
alembic
openai
azure-ai-openai
python-docx
reportlab
openpyxl
langchain
langchain-openai
tiktoken
